pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "noethersolve"
version = "1.23.0"
description = "230+ verified tools for AI agents — physics, math, genetics, enzyme kinetics, quantum mechanics, pharmacokinetics, drug interactions, epidemiology, climate physics, turbulence, topological phases, ergodic theory, optimization convergence, numerical PDEs, MHD conservation, GR constraints, seismic waves, plasma adiabatic invariants, organic chemistry, elliptic curves, intersection theory, information theory, autonomy analysis, metacognition, LLM science, networking, OS, crypto, finance. 220+ calculators + 10 lookup tables. pip install, add to Claude, instant smarter agent."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["mcp", "ai-tools", "llm", "conservation-laws", "scientific-discovery", "model-context-protocol"]

dependencies = [
    "numpy>=1.24",
    "scipy>=1.11",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
mlx = [
    "mlx>=0.18",
    "mlx-lm>=0.18",
]
torch = [
    "torch>=2.1",
    "transformers>=4.40",
    "accelerate>=0.28",
]
auto = [
    "anthropic>=0.40",
]
mcp = [
    "mcp>=1.0",
]
dashboard = [
    "matplotlib>=3.7",
]
dev = [
    "pytest>=8.0",
    "black>=24.0",
    "ruff>=0.4",
]

[project.scripts]
noethersolve-mcp = "noethersolve.mcp_server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["noethersolve*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E501"]
